+++ dashboard/src/components/partials/TableToolbarCell/FloatingTableToolbarCell.js
@@ -1,11 +1,21 @@
 import React from "react";
 import PropTypes from "prop-types";
 
+import { withStyles } from "@material-ui/core/styles";
 import Media from "react-media";
 import TableToolbarCell from "./TableToolbarCell";
 
+const styles = theme => ({
+  soft: {
+    color: theme.palette.text.secondary,
+  },
+});
+
 class FloatingTableToolbarCell extends React.Component {
   static propTypes = {
+    // the linter isn't smart enough to see we use this
+    // eslint-disable-next-line
+    classes: PropTypes.object.isRequired,
     disabled: PropTypes.bool,
     hovered: PropTypes.bool,
     children: PropTypes.func.isRequired,
@@ -21,7 +31,11 @@ class FloatingTableToolbarCell extends React.Component {
     const disabled = disabledProp || (canHover && !hovered);
 
     return (
-      <TableToolbarCell floating={canHover} disabled={disabled}>
+      <TableToolbarCell
+        className={props.classes.soft}
+        floating={canHover}
+        disabled={disabled}
+      >
         {children}
       </TableToolbarCell>
     );
@@ -36,4 +50,4 @@ class FloatingTableToolbarCell extends React.Component {
   }
 }
 
-export default FloatingTableToolbarCell;
+export default withStyles(styles)(FloatingTableToolbarCell);