+++ CHANGELOG.md
@@ -24,6 +24,8 @@ An error is now logged instead, and extraction continues after a bad line is enc
 - Keepalive events will now continue to fire after cluster restarts.
 - Fixed a panic in the dashboardd shutdown routine.
 - Fixed a bug where deleting a non-existent entity with sensuctl would not return an error.
+- Web UI - toolbar menu buttons now switch with dark theme
+- Web UI - some buttons easier to see with dark theme
 
 ### Changed
 - Improved logging for errors in proxy check requests.